Rename folder component
